# swagger.yaml
openapi: 3.0.0
info:
  title: Saint-Pierre API
  version: 1.0.0
  description: API complète pour la gestion des devoirs, cours et messagerie scolaire
  contact:
    name: Support
  x-logo:
    url: https://your-api-logo.png
servers:
  - url: http://localhost:3000
    description: Serveur de développement
  - url: https://api.saint-pierre.com
    description: Serveur de production
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token from /auth/sign-in
  schemas:
...
    Utilisateur:
      type: object
      properties:
        id_user:
          type: string
          pattern: '^[0-9]+$'
          example: '1'
        nom:
          type: string
          maxLength: 50
        prenom:
          type: string
          maxLength: 50
        login:
          type: string
          maxLength: 50
        email:
          type: string
          format: email
        role:
          $ref: '#/components/schemas/Role'
        created_at:
          type: string
          format: date-time
        eleve:
          $ref: '#/components/schemas/Eleve'
        professeur:
          $ref: '#/components/schemas/Professeur'
    Role:
      type: string
      enum: [eleve, professeur, administrateur]
    Eleve:
      type: object
      properties:
        id_user:
          type: string
          pattern: '^[0-9]+$'
        id_classe:
          type: string
          pattern: '^[0-9]+$'
          nullable: true
        annee:
          type: string
          nullable: true
        classe:
          $ref: '#/components/schemas/Classe'
        specialites:
          type: array
          items:
            $ref: '#/components/schemas/Specialite'
        options:
          type: array
          items:
            $ref: '#/components/schemas/Option'
    Professeur:
      type: object
      properties:
        id_user:
          type: string
          pattern: '^[0-9]+$'
        matiere:
          type: string
          maxLength: 50
          nullable: true
        classes_enseignees:
          type: array
          items:
            $ref: '#/components/schemas/Classe'
        specialites_enseignees:
          type: array
          items:
            $ref: '#/components/schemas/Specialite'
        options_enseignees:
          type: array
          items:
            $ref: '#/components/schemas/Option'

    # ========== CLASSE & STRUCTURES SCOLAIRES ==========
    Classe:
      type: object
      properties:
        id_classe:
          type: string
          pattern: '^[0-9]+$'
          example: '1'
        nom_classe:
          type: string
          maxLength: 50
          example: 'Terminale A'
        niveau:
          type: string
          maxLength: 20
          example: 'Terminale'
        lettre:
          type: string
          maxLength: 1
          example: 'A'
        created_at:
          type: string
          format: date-time
...
    Specialite:
      type: object
      properties:
        id_specialite:
          type: string
          pattern: '^[0-9]+$'
        nom_specialite:
          type: string
          maxLength: 50
          example: 'NSI'
        created_at:
          type: string
          format: date-time
    Option:
      type: object
      properties:
        id_option:
          type: string
          pattern: '^[0-9]+$'
        nom_option:
          type: string
          maxLength: 50
          example: 'Latin'
        created_at:
          type: string
          format: date-time

    # ========== COURS ==========
...
    Error400:
      type: object
      properties:
        error:
          type: string
          example: 'Paramètres invalides'
    Error401:
      type: object
      properties:
        error:
          type: string
          example: 'Accès refusé: token manquant'
    Error403:
      type: object
      properties:
        error:
          type: string
          example: 'Accès refusé: rôle insuffisant'
    Error404:
      type: object
      properties:
        error:
          type: string
          example: 'Ressource non trouvée'
...
    SuccessResponse:
      type: object
      properties:
        message:
          type: string
          example: 'Opération réussie'
...
security:
  - BearerAuth: []
paths:
...
  /api/users:
    get:
      summary: Lister tous les utilisateurs (Admin uniquement)
      tags: [Users]
      operationId: getAllUsers
      x-roles: [administrateur]
      responses:
        '200':
          description: Liste des utilisateurs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Utilisateur'
        '401':
          description: Non authentifié
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error401'
        '403':
          description: Rôle insuffisant
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error403'

    post:
      summary: Créer un nouvel utilisateur (Admin uniquement)
      tags: [Users]
      operationId: createUser
      x-roles: [administrateur]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                nom:
                  type: string
                  maxLength: 50
                prenom:
                  type: string
                  maxLength: 50
                email:
                  type: string
                  format: email
                password:
                  type: string
                  minLength: 6
                role:
                  $ref: '#/components/schemas/Role'
                id_classe:
                  type: string
                  pattern: '^[0-9+$'
                  nullable: true
                annee:
                  type: string
                  nullable: true
                specialites:
                  type: array
                  items:
                    type: string
                    pattern: '^[0-9+$'
                options:
                  type: array
                  items:
                    type: string
                    pattern: '^[0-9]+$'
              required: [nom, prenom, email, password]
      responses:
        '201':
          description: Utilisateur créé
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Utilisateur'
        '400':
          description: Données invalides
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error400'
        '403':
          description: Rôle insuffisant
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error403'
  /api/users/{id}:
    get:
      summary: Récupérer un utilisateur par ID (Admin uniquement)
      tags: [Users]
      operationId: getUser
      x-roles: [administrateur]
      parameters:
        - name: id
          in: path
          required: true
          description: ID de l'utilisateur
          schema:
            type: string
            pattern: '^[0-9]+$'
      responses:
        '200':
          description: Utilisateur trouvé
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Utilisateur'
        '404':
          description: Utilisateur non trouvé
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error404'
        '403':
          description: Rôle insuffisant
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error403'

    put:
      summary: Mettre à jour un utilisateur (Admin uniquement)
      tags: [Users]
      operationId: updateUser
      x-roles: [administrateur]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            pattern: '^[0-9]+$'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                nom:
                  type: string
                prenom:
                  type: string
                email:
                  type: string
                  format: email
                password:
                  type: string
                id_classe:
                  type: string
                  pattern: '^[0-9]+$'
                  nullable: true
                annee:
                  type: string
                specialites:
                  type: array
                  items:
                    type: string
                    pattern: '^[0-9]+$'
                options:
                  type: array
                  items:
                    type: string
                    pattern: '^[0-9]+$'
      responses:
        '200':
          description: Utilisateur mis à jour
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '404':
          description: Utilisateur non trouvé
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error404'
        '403':
          description: Rôle insuffisant
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error403'

    delete:
      summary: Supprimer un utilisateur (Admin uniquement)
      tags: [Users]
      operationId: deleteUser
      x-roles: [administrateur]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            pattern: '^[0-9]+$'
      responses:
        '200':
          description: Utilisateur supprimé
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '404':
          description: Utilisateur non trouvé
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error404'
        '403':
          description: Rôle insuffisant
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error403'
  /api/users/eleves/list:
    get:
      summary: Lister tous les élèves (Admin & Professeur)
      tags: [Users]
      operationId: getStudentsList
      x-roles: [administrateur, professeur]
      responses:
        '200':
          description: Liste des élèves
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id_user:
                      type: string
                    nom:
                      type: string
                    prenom:
                      type: string
                    email:
                      type: string
                    id_classe:
                      type: string
                      nullable: true
                    annee:
                      type: string
                      nullable: true
                    specialites:
                      type: array
                      items:
                        $ref: '#/components/schemas/Specialite'
                    options:
                      type: array
                      items:
                        $ref: '#/components/schemas/Option'
        '403':
          description: Rôle insuffisant
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error403'

  # ========== MATIERES ==========
...